+++ DbAccessor/src/Posts/CreatePost.cs
@@ -5,7 +5,7 @@ using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
 using MySql.Data.MySqlClient;
 
-namespace GoRideShare
+namespace GoRideShare.posts
 {
     // This class handles making a new Post
     public class CreatePost(ILogger<CreatePost> logger)
