+++ DbAccessor/src/Users/RequestObjects.cs
@@ -1,0 +1,34 @@
+using System.Text.Json.Serialization;
+
+namespace GoRideShare.users
+{
+    public class LoginCredentials(string email, string passwordHash)
+    {
+        [JsonPropertyName("email")]
+        public string Email { get; set; } = email;
+        [JsonPropertyName("password")]
+        public string PasswordHash { get; set; } = passwordHash;
+    }
+
+    public class UserRegistrationInfo
+    {
+        [JsonPropertyName("email")]
+        public string? Email { get; set; }
+
+        [JsonPropertyName("password")]
+        public string? PasswordHash { get; set; }
+
+        [JsonPropertyName("name")]
+        public string? Name { get; set; }
+
+        [JsonPropertyName("bio")]
+        public string? Bio { get; set; }
+
+        [JsonPropertyName("phone")]
+        public string? PhoneNumber { get; set; }
+
+        [JsonPropertyName("photo")]
+        public string? Photo { get; set; }
+    }
+
+}
